+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Yellowbox Statsd Changelog
+## Next
+### Added
+* `values`, `min`, `max` methods to gauge captures
 ## 0.1.1
 ### Added
 * "host" parameter to control the hostname to bind to